試驗對基本物件抽出來
