+++ 試驗/基本物件/Test字單元試驗.py
@@ -3,7 +3,6 @@ from 臺灣言語工具.解析整理.拆文分析器 import 拆文分析器
 from 臺灣言語工具.基本物件.字 import 字
 from 臺灣言語工具.基本物件.公用變數 import 無音
 from 臺灣言語工具.解析整理.型態錯誤 import 型態錯誤
-from 臺灣言語工具.音標系統.閩南語綜合標音 import 閩南語綜合標音
 
 
 class 字單元試驗(TestCase):
@@ -44,9 +43,3 @@ class 字單元試驗(TestCase):
         self.assertEqual(字物件.看音(), '')
         self.assertEqual(字物件.看分詞(), '媠')
 
-    def test_綜合標音(self):
-        字物件 = 拆文分析器.對齊字物件('意', 'i2')
-        self.assertEqual(字物件.綜合標音(閩南語綜合標音), [{
-            "漢字": "意", "臺羅數字調": "i2", "臺羅閏號調": "í",
-            "通用數字調": "i4", "吳守禮方音": "ㄧˋ"
-        }])
