+++ 試驗/基本物件/Test章單元試驗.py
@@ -3,8 +3,6 @@ from 臺灣言語工具.解析整理.拆文分析器 import 拆文分析器
 from 臺灣言語工具.基本物件.公用變數 import 分字符號
 from 臺灣言語工具.解析整理.型態錯誤 import 型態錯誤
 from 臺灣言語工具.基本物件.章 import 章
-from 臺灣言語工具.音標系統.閩南語綜合標音 import 閩南語綜合標音
-from unittest.mock import patch
 
 
 class 章單元試驗(TestCase):
@@ -75,27 +73,3 @@ class 章單元試驗(TestCase):
                          'tsiah8-pa2 0bue7 ?\ntsiah8-pa2 0ah4 !')
         self.assertEqual(章物件.看分詞(物件分句符號='\n'),
                          '食-飽｜tsiah8-pa2 未｜0bue7 ？｜?\n食-飽｜tsiah8-pa2 矣｜0ah4 ！｜!')
-
-    def test_綜合標音章物件(self):
-        章物件 = 拆文分析器.對齊章物件('點仔膠，黏著跤，', 'tiam2-a2-ka1, liam5-tioh8 kha1,')
-        頭一句, 上尾句 = 章物件.綜合標音(閩南語綜合標音)
-        self.assertIn('漢字', 頭一句)
-        self.assertIn('臺羅數字調', 頭一句)
-        self.assertIn('臺羅閏號調', 頭一句)
-        self.assertIn('通用數字調', 頭一句)
-        self.assertIn('吳守禮方音', 頭一句)
-        self.assertIn('分詞', 頭一句)
-        self.assertIn('漢字', 上尾句)
-        self.assertIn('臺羅數字調', 上尾句)
-        self.assertIn('臺羅閏號調', 上尾句)
-        self.assertIn('通用數字調', 上尾句)
-        self.assertIn('吳守禮方音', 上尾句)
-        self.assertIn('分詞', 上尾句)
-
-    @patch('臺灣言語工具.基本物件.句.句.綜合標音')
-    def test_綜合標音用句的來鬥(self, 句綜合標音mock):
-        章物件 = 拆文分析器.對齊章物件('點仔膠，黏著跤，', 'tiam2-a2-ka1, liam5-tioh8 kha1,')
-        self.assertEqual(章物件.綜合標音(閩南語綜合標音), [
-            句綜合標音mock.return_value[0],
-            句綜合標音mock.return_value[0],
-        ])
