+++ 試驗/基本物件/Test組單元試驗.py
@@ -2,7 +2,6 @@ from unittest.case import TestCase
 from 臺灣言語工具.解析整理.拆文分析器 import 拆文分析器
 from 臺灣言語工具.解析整理.型態錯誤 import 型態錯誤
 from 臺灣言語工具.基本物件.組 import 組
-from 臺灣言語工具.音標系統.閩南語綜合標音 import 閩南語綜合標音
 
 
 class 組單元試驗(TestCase):
@@ -62,30 +61,3 @@ class 組單元試驗(TestCase):
             公鄉所組物件.看分詞(),
             '鄉-公-所｜-kong1- 鄉-公-所｜hiang1--soo2'
         )
-
-    def test_一組綜合標音json格式(self):
-        組物件 = 拆文分析器.對齊組物件('椅仔', 'i2-a2')
-        頭一个詞, = 組物件.綜合標音(閩南語綜合標音)
-        self.assertIn('漢字', 頭一个詞)
-        self.assertIn('臺羅數字調', 頭一个詞)
-        self.assertIn('臺羅閏號調', 頭一个詞)
-        self.assertIn('通用數字調', 頭一个詞)
-        self.assertIn('吳守禮方音', 頭一个詞)
-        self.assertIn('分詞', 頭一个詞)
-
-    def test_一組兩詞綜合標音json格式(self):
-        組物件 = 拆文分析器.對齊組物件('椅仔', 'i2 a2')
-        頭一个詞, = 組物件.綜合標音(閩南語綜合標音)
-        self.assertIn('漢字', 頭一个詞)
-        self.assertIn('臺羅數字調', 頭一个詞)
-        self.assertIn('臺羅閏號調', 頭一个詞)
-        self.assertIn('通用數字調', 頭一个詞)
-        self.assertIn('吳守禮方音', 頭一个詞)
-        self.assertIn('分詞', 頭一个詞)
-
-    def test_空組綜合標音莫例外(self):
-        # 因為攏用佇輸出，愛檢查空愛佇程式別位檢查
-        組物件 = 組()
-        self.assertEqual(組物件.綜合標音(閩南語綜合標音), [
-            {}
-        ])
