+++ 試驗/基本物件/Test詞單元試驗.py
@@ -4,7 +4,6 @@ from 臺灣言語工具.基本物件.公用變數 import 分字符號
 from 臺灣言語工具.基本物件.字 import 字
 from 臺灣言語工具.基本物件.詞 import 詞
 from 臺灣言語工具.解析整理.型態錯誤 import 型態錯誤
-from 臺灣言語工具.音標系統.閩南語綜合標音 import 閩南語綜合標音
 
 
 class 詞單元試驗(TestCase):
@@ -67,20 +66,3 @@ class 詞單元試驗(TestCase):
         ]
         鄉所分詞答案 = '鄉-公-所｜hiang1--soo2'
         self.assertEqual(鄉所詞物件.看分詞(), 鄉所分詞答案)
-
-    def test_綜合標音json格式(self):
-        詞物件 = 拆文分析器.對齊詞物件('椅仔', 'i2-a2')
-        頭一个詞, = 詞物件.綜合標音(閩南語綜合標音)
-        self.assertIn('漢字', 頭一个詞)
-        self.assertIn('臺羅數字調', 頭一个詞)
-        self.assertIn('臺羅閏號調', 頭一个詞)
-        self.assertIn('通用數字調', 頭一个詞)
-        self.assertIn('吳守禮方音', 頭一个詞)
-        self.assertIn('分詞', 頭一个詞)
-
-    def test_綜合標音空詞莫例外(self):
-        # 因為攏用佇輸出，愛檢查空愛佇程式別位檢查
-        詞物件 = 拆文分析器.建立詞物件('')
-        self.assertEqual(詞物件.綜合標音(閩南語綜合標音), [
-            {}
-        ])
